refactor: moge methods to new file
